pyproject.toml
[tool.poetry]
package-mode = false

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
pyyaml = "^6.0.2"
torchvision = "0.17.2"
torch = "2.2.2"
numpy = "1.25.2"
torchaudio = "2.2.2"
matplotlib = "^3.9.2"
scipy = "^1.14.1"
lightning = "^2.4.0"
pysteps = "^1.11.0"
tqdm = "4.64"
torchcfm = "^1.0.5"
torchdiffeq = "^0.2.4"
jaxtyping = "0.2.19"
gpytorch = "1.11.0"
properscoring = "^0.1"
wandb = "^0.17.9"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
